feat(plats): implement Twitch platform
Implement the interface `Platform` for Twitch.
Add an empty extension for browser to load.
Add the flag `--disable-blink-features=AutomationControlled` and load extension (`src/utils/extension`) to browser to bypass browser check.
By the way, Twitch is so annoying that I almost rewrote the entire program's logic. And fortunately I didn't.